refactor: Delete job by ID in deleteJob function
This commit updates the deleteJob function in the jobs.controller.js file to delete a job from the MongoDB database based on its ID. It first retrieves the job using the Job.findById method and throws a 404 error if no job is found. Then, it uses the job.deleteOne method to remove the job from the database. Finally, it returns a JSON response with a success message indicating that the job has been deleted. This change improves the functionality and reliability of the delete job operation.