refactor: Update updateJob function to include location field in job updates
This commit modifies the updateJob function in the jobs.controller.js file to include the location field in the updates made to a job. Previously, only the company and position fields were updated. Now, if a location field is provided in the request body, it will be used to update the job's location. This change enhances the functionality of the update job operation by allowing for more comprehensive updates to job details.