feat: Generate JWT token on successful login
The code changes in the `auth.controller.js` file add the functionality to generate a JSON Web Token (JWT) upon successful login. This token is created using the `createToken` function from the `jwt.utils.js` file, which takes the user's ID and role as the payload. The generated token is then returned in the response body. This feature enhances the authentication process by providing a secure and convenient way to authenticate subsequent requests.