refactor: Add authenticateUser middleware for authentication
The code changes in the auth.middleware.js file add a new middleware function called authenticateUser. This middleware checks if the request contains a token in the cookies and verifies the token using the verifyToken function from the jwt.utils.js file. If the token is valid, it sets the user ID and role in the request object. If the token is invalid, it throws an UnauthenticatedError. This refactor enhances the security of the API by ensuring that only authenticated users can access certain routes.