refactor: Update jwt.utils.js to include verifyToken function
The code changes in the jwt.utils.js file add a new function called verifyToken. This function takes a JWT token as input and verifies its authenticity using the JWT_SECRET_KEY from the environment variables. The decoded token is returned if it is valid. This refactor enhances the security of the API by providing a centralized and reusable function for token verification.