refactor: Update authenticateUser middleware to set user object in request
The code changes in the auth.middleware.js file refactor the authenticateUser middleware to set the user object in the request. Instead of setting the id and role separately, the user object now contains the id and role properties. This refactor improves the readability and consistency of the code by encapsulating the user information in a single object.