refactor: Update logout function to clear token cookie
The code changes in the auth.controller.js file refactor the logout function to clear the token cookie. Instead of setting the token cookie to 'logout', the cookie is now cleared by setting the expires option to a past date. This refactor improves the security of the logout functionality by ensuring that the token is properly invalidated and cannot be used for further authentication.