feat: Add user route for retrieving user information
The code changes in the `user.route.js` file add a new route `/api/v1/users` to the API. This route is responsible for retrieving user information and is handled by the `getUser` controller function. This feature enhances the functionality of the API by providing an endpoint specifically for retrieving user data.