feat: Add updateUser controller function for updating user profile
This code change adds a new controller function, updateUser, to the user.controller.js file. The function is responsible for updating the user's profile and is accessed via the PATCH method at the /api/v1/users endpoint. This feature enhances the functionality of the API by providing an endpoint for users to update their profile information.