feat: Add getUser controller function for retrieving user profile
This code change adds a new controller function, getUser, to the user.controller.js file. The function is responsible for retrieving the user's profile and is accessed via the GET method at the /api/v1/users endpoint. This feature enhances the functionality of the API by providing an endpoint for users to retrieve their profile information.