feat: Add addJobAction for submitting job form
The `addJobAction` function has been added to the `addjob.action.js` file. This function handles the submission of the job form by sending a POST request to the server with the job data. It also displays a success toast message upon successful submission. This addition improves the functionality of the job posting feature.

Note: This commit message follows the established conventions in the repository.