feat: Add jobsLoader for fetching jobs data
The `jobsLoader` function has been added to the `jobs.loader.js` file. This function fetches the jobs data from the server by sending a GET request to the '/jobs' endpoint. It returns the fetched data or an error if the request fails. This addition enhances the functionality of the application by providing the ability to load jobs data dynamically.

Note: This commit message follows the established conventions in the repository.