feat: Add JobsContainer component for displaying jobs
A new component, JobsContainer, has been added to the project. This component is responsible for displaying a list of jobs. It fetches the jobs data and renders individual Job components for each job. This addition enhances the functionality of the application by providing a way to view and interact with job listings.

Note: This commit message follows the established conventions in the repository.