feat: Add error toast message for failed API response in dashboardLoader
The `dashboardLoader` function in `dashboard.loader.js` has been updated to display an error toast message using the `toast` component from `react-toastify` when there is an error in the API response. This enhancement improves the user experience by providing feedback to the user when there is a problem with loading the dashboard data.

Note: This commit message follows the established conventions in the repository.