feat: Add editJobLoader for fetching job data for editing
A new loader, editJobLoader, has been added to fetch job data for editing. This loader is responsible for making an API request to retrieve the job data based on the provided ID. If successful, the data is returned. Otherwise, an error toast message is displayed and the user is redirected to the dashboard. This addition enhances the functionality of the application by providing a way to fetch and display job data for editing.

Note: This commit message follows the established conventions in the repository.