feat: Add editJobAction for updating job details
The editJobAction function has been added to the project. This function is responsible for updating the details of a job, such as the company, position, status, type, and location. It uses the customFetch utility to make a PATCH request to the server with the updated data. If the request is successful, a success toast message is displayed and the user is redirected to the dashboard. If there is an error, an error toast message is displayed. This addition enhances the functionality of the application by providing a way to modify existing job listings.

Note: This commit message follows the established conventions in the repository.