feat: Add deleteJobLoader for deleting job postings
A new loader, deleteJobLoader, has been added to handle the deletion of job postings. This loader makes a DELETE request to the server to delete the job with the specified ID. If the deletion is successful, a success toast message is displayed and the user is redirected to the jobs dashboard. If there is an error, an error toast message is displayed. This addition enhances the functionality of the application by providing a way to delete job postings.

Note: This commit message follows the established conventions in the repository.