feat: Add adminLoader for fetching admin data
A new loader, adminLoader, has been added to fetch admin data from the server. This loader makes a GET request to '/admin/stats' and returns the data. If the request fails, an error toast message is displayed and the user is redirected to the jobs dashboard. This addition enhances the functionality of the application by providing a way to load admin data.

Note: This commit message follows the established conventions in the repository.