refactor: Update NavLinks component to conditionally render admin link
The NavLinks component has been updated to conditionally render the admin link based on the user's role. If the user is an admin, the admin link will be displayed in the navigation bar. This refactor improves the user experience by providing a more tailored navigation menu for admins.

Note: This commit message follows the established conventions in the repository.