feat: Add avatar image support in user dropdown
This commit adds support for displaying an avatar image in the user dropdown menu. If the user has uploaded an avatar image, it will be displayed instead of the default user icon. This enhancement improves the visual representation of the user in the dropdown menu.

Note: This commit message follows the established conventions in the repository.